peridot: parts: Introduce Display Satutaion Service and TileUtils
* Adjust Saturation Value Handling for Device Color Mode Compatibility

This update refines the saturation adjustment mechanism. When the seek bar progress is set, the method executes
the command "service call SurfaceFlinger 1022 f [saturation]" via Runtime.getRuntime().exec() to apply the saturation changes.
However, it was observed that when the device's color mode is set to "saturated," a saturation value of 1.0 does not trigger any visual changes.
To address this, the code now sets a value of 1.001 for a seek bar value of 100, ensuring the desired effect is consistently applied.

* TileUtils Implementation: Implemented TileUtils to handle the addition of Quick Settings tiles, specifically for the saturation feature.
This is done with reference to the guidelines provided by Google
https://developer.android.com/develop/ui/views/quicksettings-tiles#prompt-user
this ensuring a seamless integration and user experience when creating new tiles.

Signed-off-by: kenway214 <kenway214@outlook.com>

peridot: parts: Rewrite in kotlin

Signed-off-by: kenway214 <kenway214@outlook.com>
Co-Authored-By: AnierinB <anierin@evolution-x.org>